# jagatrakshak_openapi.json
{
    "openapi":"3.0.3",
    "info":{
       "title":"FastAPI",
       "version":"0.1.0"
    },
    "servers": [
      {
        "url": "https://jagatrakshak.loca.lt/api"
      }
    ],
    "paths":{
      "/notify/sms/":{
         "post":{
            "summary":"Send SMS Notification",
            "description":"Send an SMS notification to a specified phone number.",
            "operationId":"send_sms_notify_sms_post",
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/SMSRequest"
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "type":"string",
                           "title":"Response Send SMS Notify Sms Post"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/notify/email/":{
         "post":{
            "summary":"Send Email Notification",
            "description":"Send an email notification to a specified email address with a custom message body and subject.",
            "operationId":"send_email_notify_email_post",
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/EmailRequest"
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "type":"string",
                           "title":"Response Send Email Notify Email Post"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/notify/call/":{
         "post":{
            "summary":"Make Call Notification",
            "description":"Make a call notification to a specified phone number.",
            "operationId":"make_call_notify_call_post",
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/CallRequest"
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "type":"string",
                           "title":"Response Make Call Notify Call Post"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
...
      "/users/search/volunteers": {
         "get": {
            "summary": "Search for volunteer users",
            "description": "Search for volunteer users based on various criteria.",
            "operationId": "search_volunteer_users_get",
            "parameters": [
               {
                  "name": "country",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "string",
                     "description": "Country to filter by"
                  }
               },
               {
                  "name": "state",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "string",
                     "description": "State to filter by"
                  }
               },
               {
                  "name": "city",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "string",
                     "description": "City to filter by"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "$ref": "#/components/schemas/UserResponse"
                           }
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/users/{email_id}": {
         "get": {
            "summary": "Get user by email ID",
            "description": "Retrieve user information using their email ID.",
            "operationId": "get_user_by_email_id",
            "parameters": [
               {
                  "name": "email_id",
                  "in": "path",
                  "required": true,
                  "schema": {
                     "type": "string"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UserResponse"
                        }
                     }
                  }
               },
               "404": {
                  "description": "User Not Found",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/users/by_id/{user_id}": {
         "get": {
            "summary": "Get user by ID",
            "description": "Retrieve user information using their ID.",
            "operationId": "get_user_by_id",
            "parameters": [
               {
                  "name": "user_id",
                  "in": "path",
                  "required": true,
                  "schema": {
                     "type": "integer"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UserResponse"
                        }
                     }
                  }
               },
               "404": {
                  "description": "User Not Found",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPError"
                        }
                     }
                  }
               }
            }
         }
      },
...
      "/incidents/user/{user_email}": {
         "get": {
            "summary": "Get incidents by user email",
            "description": "Retrieve all incidents reported by a specific user.",
            "operationId": "get_incidents_by_user_email",
            "parameters": [
               {
                  "name": "user_email",
                  "in": "path",
                  "required": true,
                  "schema": {
                     "type": "string"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "$ref": "#/components/schemas/IncidentResponse"
                           }
                        }
                     }
                  }
               },
               "404": {
                  "description": "User Not Found",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPError"
                        }
                     }
                  }
               }
            }
         }
      },
...
    },
    "components":{
       "schemas":{
...
         "IncidentResponse": {
            "properties": {
               "incident_id": {
                  "type": "integer",
                  "title": "Incident ID"
               },
               "user_email": {
                  "type": "string",
                  "title": "User Email"
               },
               "created_date": {
                  "type": "string",
                  "format": "date-time",
                  "title": "Created Date"
               },
               "updated_date": {
                  "type": "string",
                  "format": "date-time",
                  "title": "Updated Date"
               },
               "incident_place": {
                  "type": "string",
                  "title": "Incident Place"
               },
               "incident_state": {
                  "type": "string",
                  "title": "Incident State"
               },
               "incident_country": {
                  "type": "string",
                  "title": "Incident Country"
               },
               "incident_city": {
                  "type": "string",
                  "title": "Incident City"
               },
               "incident_type": {
                  "type": "string",
                  "title": "Incident Type"
               },
               "incident_description": {
                  "type": "string",
                  "title": "Incident Description"
               },
               "victim_type": {
                  "type": "string",
                  "title": "Victim Type"
               },
               "victim_details": {
                  "type": "string",
                  "title": "Victim Details"
               },
               "incident_status": {
                  "type": "string",
                  "title": "Incident Status"
               },
               "comment": {
                  "type": "string",
                  "title": "Comment"
               }
            },
            "type": "object",
            "required": [
               "incident_id",
               "user_email",
               "created_date",
               "updated_date",
               "incident_place",
               "incident_state",
               "incident_country",
               "incident_city",
               "incident_type",
               "victim_type",
               "incident_status"
            ],
            "title": "IncidentResponse"
         },
         "UserResponse":{
             "properties":{
                "id":{
                   "type":"integer",
                   "title":"User Id"
                },
                "name":{
                   "type":"string",
                   "title":"Name"
                },
                "email":{
                   "type":"string",
                   "title":"Email"
                },
                "city":{
                   "type":"string",
                   "title":"City"
                },
                "phone":{
                   "type":"string",
                   "title":"Phone"
                }
             },
             "type":"object",
             "required":[
                "id",
                "name",
                "email",
                "city",
                "phone"
             ],
             "title":"UserResponse"
          },
         "SMSRequest":{
            "properties":{
               "phone_number":{
                  "type":"string",
                  "title":"Phone Number",
                  "example":"9876543210"
               },
               "message":{
                  "type":"string",
                  "title":"Message",
                  "example":"This is a notification message from JagatRakshak Emergency Response System."
               }
            },
            "type":"object",
            "required":[
               "phone_number",
               "message"
            ],
            "title":"SMSRequest"
         },
         "EmailRequest":{
            "properties":{
               "to":{
                  "type":"string",
                  "title":"Email",
                  "example":"user@example.com"
               },
               "subject":{
                  "type":"string",
                  "title":"Subject",
                  "example":"Urgent: Action Required"
               },
               "body":{
                  "type":"string",
                  "title":"Body",
                  "example":"This is a notification email from JagatRakshak Emergency Response System."
               }
            },
            "type":"object",
            "required":[
               "to",
               "subject",
               "body"
            ],
            "title":"EmailRequest"
         },
         "CallRequest":{
            "properties":{
               "phone_number":{
                  "type":"string",
                  "title":"Phone Number",
                  "example":"9876543210"
               },
               "message":{
                  "type":"string",
                  "title":"Message",
                  "example":"This is a notification call from JagatRakshak Emergency Response System."
               }
            },
            "type":"object",
            "required":[
               "phone_number",
               "message"
            ],
            "title":"CallRequest"
         },
...
         "ValidationError":{
            "properties":{
               "loc":{
                  "items":{
                     "type":"string"
                  },
                  "type":"array",
                  "title":"Location"
               },
               "msg":{
                  "type":"string",
                  "title":"Message"
               },
               "type":{
                  "type":"string",
                  "title":"Error Type"
               }
            },
            "type":"object",
            "required":[
               "loc",
               "msg",
               "type"
            ],
            "title":"ValidationError"
         },
         "HTTPError":{
            "properties":{
               "detail":{
                  "type":"string",
                  "title":"Detail",
                  "example":"User not found"
               }
            },
            "type":"object",
            "title":"HTTPError"
         },
         "HTTPValidationError":{
            "properties":{
               "detail":{
                  "items":{
                     "$ref":"#/components/schemas/ValidationError"
                  },
                  "type":"array",
                  "title":"Detail"
               }
            },
            "type":"object",
            "title":"HTTPValidationError"
         }
       }
    }
 }